rename default_allow module to defaults
